대공사

+++ src/main/java/com/green/muziuniv_be_notuser/entity/application/Application.java
@@ -21,7 +21,7 @@ public class Application extends CreatedAt {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "app_id")
-    private int appId;
+    private Long appId;
 
     @Embedded
     private UserId userId;
